pyproject.toml
[tool.poetry]
name = "2024-p2"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^3.0.2"
flask-sqlalchemy = "^3.1.1"
loguru = "^0.7.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"